1.4 anddroid

+++ lib/request/mine_api.dart
@@ -14,10 +14,12 @@ import 'package:yimareport/config/project_config.dart';
 import 'package:yimareport/db/entities/local_record.dart';
 import 'package:yimareport/db/entities/member_record.dart';
 import 'package:yimareport/entities/cycle_entity.dart';
+import 'package:yimareport/entities/is_login_entity.dart';
 import 'package:yimareport/entities/login_entity.dart';
 import 'package:yimareport/entities/login_sync_entity.dart';
 import 'package:yimareport/entities/tag_sync_entity.dart';
 import 'package:yimareport/entities/version_entity.dart';
+import 'package:yimareport/request/mark_api.dart';
 import 'package:yimareport/utils/dialog.dart';
 import 'package:yimareport/utils/toast_util.dart';
 
@@ -49,6 +51,12 @@ class MineAPI {
   final LOGIN = "/wechat/login";
   Future<LoginEntity> login(Map<String, dynamic> params,{BuildContext? buildContext}) async {
     LoginEntity entity = await HttpManager.instance.request<LoginEntity>(Method.GET, LOGIN, params: params, buildContext: buildContext, contentType: HttpContentType.form);
+    return entity;
+  }
+  // 登录验证
+  final ISLOGIN = "/wechat/is-login";
+  Future<IsLoginEntity> isLogin(Map<String, dynamic> params,{BuildContext? buildContext}) async {
+    IsLoginEntity entity = await HttpManager.instance.request<IsLoginEntity>(Method.GET, ISLOGIN, params: params, buildContext: buildContext, contentType: HttpContentType.form);
     return entity;
   }
   // 退出登录
@@ -100,6 +108,15 @@ class MineAPI {
       "type": record.type
     };
  }
+  _recordToMapWithId(record) {
+    if(record == null) return null;
+    return {
+      "markAt": record.markAt,
+      "createAt": record.createAt,
+      "type": record.type,
+      "id": record.id
+    };
+  }
  getUUID() async {
     // return "36a6a32e32ce65de";
    String uuid, dev_info;
@@ -115,8 +132,35 @@ class MineAPI {
    }
    return uuid;
  }
+ //登录校验 如果没登录则返回true
+  checkLogin() async{
+    var account = getAccount();
+    SharedPreferences sharedPreferences = await SharedPreferences.getInstance();
+    String lastUserId = sharedPreferences.getString("LASTUSERID") ?? '';
+    String userId = account?.user_id ?? lastUserId;
+    var connectivityResult = await (Connectivity().checkConnectivity());
+    if (connectivityResult == ConnectivityResult.mobile || connectivityResult == ConnectivityResult.wifi) {
+      if(userId != '') {
+        IsLoginEntity res =  await isLogin({"uuid": await getUUID(), "user_id": userId,});
+        var result = res.data.isLogin ?? false;
+        if(!result) {
+          //删本地数据
+          await DBAPI.sharedInstance.memberRecordDao.deleteAll();
+          await DBAPI.sharedInstance.markDao.deleteAllMemberMarks();
+          //删除本地登录缓存
+          SharedPreferences sharedPreferences = await SharedPreferences.getInstance();
+          sharedPreferences.remove("LASTUSERID");
+          await clearAccount();
+        }
+        return res.data.isLogin ?? false;
+      }
+    }
+    return true;
+  }
  //同步周期数据方法
  memberSyncCircle() async{
+    var isCheck = await checkLogin();
+    if(!isCheck) return;
     var connectivityResult = await (Connectivity().checkConnectivity());
     if (connectivityResult == ConnectivityResult.mobile || connectivityResult == ConnectivityResult.wifi) {
       var account = getAccount();
@@ -176,17 +220,26 @@ class MineAPI {
     if(isRmLocData) {
       //清空脏数据表
       await DBAPI.sharedInstance.localRecordDao.deleteAll();
+      //同步野数据
+      await MarkAPI.instance.markLoginSyncData(buildContext: buildContext);
+      //清空本地野数据
+      await DBAPI.sharedInstance.markDao.deleteAllLocalMarks();
     }
+
     //测试
     // MyDialog.showAlertDialog(buildContext!, () { }, title: "调试提示", message: "服务器isRmLocData字段为${isRmLocData ? 'true. 会删除野数据' : 'false. 会保留野数据'}");
   }
-  //返回未同步代表删除的第一条数据
+  //返回未同步代表删除的后2条数据（只允许操作4天内的 2条够了）
   _findMemberNeedSyncDeleteRecord() async {
     List<MemberRecord>? needDel = await DBAPI.sharedInstance.memberRecordDao.findFirstDelRecord();
     return needDel ?? [];
   }
   //会员同步数据 1， 检查网络 2，有网 先提交数据  有操作则写本地会员表， 调接口 清空本地并写入 2， 无网 写本地
   memberSyncData({BuildContext? buildContext}) async {
+    // var account = getAccount();
+    // if(account == null) return;
+    var isCheck = await checkLogin();
+    if(!isCheck) return;
     var connectivityResult = await (Connectivity().checkConnectivity());
     if (connectivityResult == ConnectivityResult.mobile || connectivityResult == ConnectivityResult.wifi) {
       //获取本地会员数据（没有merged的）
@@ -195,20 +248,31 @@ class MineAPI {
       //判断第一条是不是删除数据
       List<MemberRecord> delItemArr = await _findMemberNeedSyncDeleteRecord();
       if(delItemArr.length > 0 && delItemArr.first.isDeleted != 0) {
-        //调删除接口
-        await periodDelSync({"user_id": delItemArr.first.memberID, "markAt": delItemArr.first.markAt, "type": delItemArr.first.type, "createAt": delItemArr.first.createAt, "uuid": await getUUID()});
+        for(int i = 0; i < delItemArr.length; i++) {
+          var delI = delItemArr[i];
+          //调删除接口
+          await periodDelSync({"user_id": delI.memberID, "markAt": delI.markAt, "type": delI.type, "createAt": delI.createAt, "uuid": await getUUID()});
+        }
+        // //调删除接口
+        // await periodDelSync({"user_id": delItemArr.first.memberID, "markAt": delItemArr.first.markAt, "type": delItemArr.first.type, "createAt": delItemArr.first.createAt, "uuid": await getUUID()});
       }
       //需要同步的数据
       List<MemberRecord> _filterData = _dataSource.where((element) => element.isDeleted == 0 && element.isLogout == 0).toList();
       //同步数据
-      var jsonArr = _filterData.map((e) => _recordToMap(e)).toList();
-      var params = {
-        "user_id": _dataSource.first.memberID,
-        "data": jsonArr,
-        "uuid": await getUUID()
-      };
-      TagSyncEntity entity =  await periodSync(params, buildContext: buildContext);
-      List<TagSyncData> serverData = entity.data;
+      List<TagSyncData>? serverData;
+      var unmergeItems = _dataSource.where((element) => element.isDeleted == 0 && element.isLogout == 0 && element.isMerged == 0).toList();
+      if(unmergeItems.length > 0) {
+        var jsonArr = _filterData.map((e) => _recordToMap(e)).toList();
+        var params = {
+          "user_id": _dataSource.first.memberID,
+          "data": jsonArr,
+          "uuid": await getUUID()
+        };
+        TagSyncEntity entity =  await periodSync(params, buildContext: buildContext);
+        serverData = entity.data;
+      }
+
+
       //判断最后一条是不是退出登录
       // MemberRecord lastItem = _dataSource.last;
       MemberRecord? lastItem = await DBAPI.sharedInstance.memberRecordDao.findLastOne();
@@ -219,17 +283,18 @@ class MineAPI {
         //清空本地会员表数据
         await DBAPI.sharedInstance.memberRecordDao.deleteAll();
       } else {
+        if(serverData == null) return;
         //接口返回数据写入本地会员表
         await DBAPI.sharedInstance.memberRecordDao.deleteAll();
         var userId = getAccount()?.user_id ?? "";
         var memberData = serverData.map((e) => MemberRecord(null, e.markAt, e.createAt, int.parse(e.type), userId, isMerged: 1)).toList();
         await DBAPI.sharedInstance.memberRecordDao.batchInsertRecords(memberData);
         //判断最后一条是不是删除， 并比较时间戳， 如果是时间最新一条的话追加上去
-        var memberLastRecord = await DBAPI.sharedInstance.memberRecordDao.findLastRecord();
-        if((lastItem?.isDeleted ?? 0) != 0 && memberLastRecord != null && int.parse(lastItem!.createAt) > int.parse(memberLastRecord.createAt)) {
-          lastItem.id = null;
-          await DBAPI.sharedInstance.memberRecordDao.insertRecord(lastItem);
-        }
+        // var memberLastRecord = await DBAPI.sharedInstance.memberRecordDao.findLastRecord();
+        // if((lastItem?.isDeleted ?? 0) != 0 && memberLastRecord != null && int.parse(lastItem!.createAt) > int.parse(memberLastRecord.createAt)) {
+        //   lastItem.id = null;
+        //   await DBAPI.sharedInstance.memberRecordDao.insertRecord(lastItem);
+        // }
       }
     } else {
       //无网不做处理
@@ -241,14 +306,20 @@ class MineAPI {
     if (connectivityResult == ConnectivityResult.mobile || connectivityResult == ConnectivityResult.wifi) {
       //同步数据
       await memberSyncData(buildContext: buildContext);
+      //同步新标记数据
+      await MarkAPI.instance.markSyncData();
       //调退出登录接口
       await logout({ "uuid": await getUUID(), "user_id": getAccount()?.user_id}, buildContext: buildContext);
       //清空本地会员表数据
       await DBAPI.sharedInstance.memberRecordDao.deleteAll();
+      //清除新标记
+      await DBAPI.sharedInstance.markDao.deleteAllMemberMarks();
     } else {
       //无网会员表写入一条退出登录记录， 等待下次同步时（memberSyncData）处理
       await DBAPI.sharedInstance.memberRecordDao.insertRecord(MemberRecord(null, "0", "0", 0, getAccount()?.user_id ?? '', isLogout: 1));
     }
+    SharedPreferences sharedPreferences = await SharedPreferences.getInstance();
+    sharedPreferences.setString("LASTUSERID", getAccount()?.user_id ?? '');
     await clearAccount();
   }
   //插入标记数据， 姨妈来了 姨妈走了
@@ -279,7 +350,34 @@ class MineAPI {
       await DBAPI.sharedInstance.localRecordDao.insertRecord(localRecord);
     }
   }
-
+  //新版的删除记录
+  newDeleteRecord({required BuildContext buildContext, required int id}) async {
+    var account = this.getAccount();
+    if(account != null) {
+      //登录状态 删除会员表
+      MemberRecord? delItem = await DBAPI.sharedInstance.memberRecordDao.findRecordById(id);
+      if(delItem != null) {
+        if(delItem.isMerged == 0) {//没merge 直接删
+          await DBAPI.sharedInstance.memberRecordDao.deleteById(id);
+        } else {
+          delItem.isDeleted = 1;
+          await DBAPI.sharedInstance.memberRecordDao.updateRecord(delItem);
+          //判断该记录是否merged， 如果merged过需要调接口
+          var connectivityResult = await (Connectivity().checkConnectivity());
+          var isConnected = connectivityResult == ConnectivityResult.mobile || connectivityResult == ConnectivityResult.wifi;
+          if( isConnected) {
+            //删远程
+            await periodDelSync({"user_id": getAccount()?.user_id!, "markAt": delItem.markAt, "type": delItem.type, "createAt": delItem.createAt, "uuid": await getUUID()});
+            //再同步
+            // await memberSyncData(buildContext: buildContext);
+          }
+        }
+      }
+    } else {
+      //脏数据表 直接删
+      await DBAPI.sharedInstance.localRecordDao.deleteById(id);
+    }
+  }
   //删除最后一条数据
   deleteLastRecord(BuildContext buildContext) async {
     var account = this.getAccount();
@@ -294,7 +392,7 @@ class MineAPI {
           MyDialog.showAlertDialog(buildContext, () {
           }, title: "没有可以删除的记录", message: "仅支持删除四天内的最后一条数据哦", isOnlySureBtn: true, sureBtnTitle: "好", sureBtnTitleColor: Colors.blue);
         } else {
-          MyDialog.showAlertDialog(buildContext, () async {
+          // await MyDialog.showAlertDialog(buildContext, () async {
             record.isDeleted = 1;
             await DBAPI.sharedInstance.memberRecordDao.updateRecord(record);
             //判断该记录是否merged， 如果merged过需要调接口
@@ -309,12 +407,12 @@ class MineAPI {
               await memberSyncData(buildContext: buildContext);
             }
             showToast("删除记录成功");
-          }, title: "删除记录？",
-              message: "${DateUtil.formatDateMs(int.parse(record.markAt),format: "yyyy年MM月dd日")}：${record.type == 1 ? "姨妈开始" : "姨妈结束"}",
-              sureBtnTitle: "删除",
-              sureBtnTitleColor: Colors.red,
-              cancelBtnTitleColor: Colors.blue
-          );
+          // }, title: "删除记录？",
+          //     message: "${DateUtil.formatDateMs(int.parse(record.markAt),format: "yyyy年MM月dd日")}：${record.type == 1 ? "姨妈开始" : "姨妈结束"}",
+          //     sureBtnTitle: "删除",
+          //     sureBtnTitleColor: Colors.red,
+          //     cancelBtnTitleColor: Colors.blue
+          // );
         }
       } else {
         MyDialog.showAlertDialog(buildContext, () {}, title: "没有可以删除的记录", message: "仅支持删除四天内的最后一条数据哦", isOnlySureBtn: true, sureBtnTitle: "好", sureBtnTitleColor: Colors.blue);
@@ -334,17 +432,17 @@ class MineAPI {
             // Navigator.of(context).pop();
           }, title: "没有可以删除的记录", message: "仅支持删除四天内的最后一条数据哦", isOnlySureBtn: true, sureBtnTitle: "好", sureBtnTitleColor: Colors.blue);
         } else {
-          MyDialog.showAlertDialog(buildContext, () async {
+          // MyDialog.showAlertDialog(buildContext, () async {
             record.isDeleted = 1;
             await DBAPI.sharedInstance.localRecordDao.updateRecord(record);
             showToast("删除记录成功");
             // Navigator.of(context).pop();
-          }, title: "删除记录？",
-              message: "${DateUtil.formatDateMs(int.parse(record.markAt),format: "yyyy年MM月dd日")}：${record.type == 1 ? "姨妈开始" : "姨妈结束"}",
-              sureBtnTitle: "删除",
-              sureBtnTitleColor: Colors.red,
-              cancelBtnTitleColor: Colors.blue
-          );
+          // }, title: "删除记录？",
+          //     message: "${DateUtil.formatDateMs(int.parse(record.markAt),format: "yyyy年MM月dd日")}：${record.type == 1 ? "姨妈开始" : "姨妈结束"}",
+          //     sureBtnTitle: "删除",
+          //     sureBtnTitleColor: Colors.red,
+          //     cancelBtnTitleColor: Colors.blue
+          // );
         }
       } else {
         MyDialog.showAlertDialog(buildContext, () {}, title: "没有可以删除的记录", message: "仅支持删除四天内的最后一条数据哦", isOnlySureBtn: true, sureBtnTitle: "好", sureBtnTitleColor: Colors.blue);
@@ -367,10 +465,10 @@ class MineAPI {
     var account = this.getAccount();
     if(account != null) {
       MemberRecord? last = await DBAPI.sharedInstance.memberRecordDao.findLastRecord();
-      return _recordToMap(last);
+      return _recordToMapWithId(last);
     } else {
       LocalRecord? last = await DBAPI.sharedInstance.localRecordDao.findLastRecord();
-      return _recordToMap(last);
+      return _recordToMapWithId(last);
     }
 
   }
